NAT/Net: Try to placate VisualC that is really paranoid about integral type sizes.  Follow up fix for r142754.  bugref:9929.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@87723 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/NetworkServices/NAT/VBoxNetLwipNAT.cpp
@@ -634,7 +634,7 @@ int VBoxNetLwipNAT::ipv4LoopbackMapInit()
     /*
      * Process mappings of the form "127.x.y.z=off"
      */
-    size_t dst = 0;
+    unsigned int dst = 0;      /* typeof(ip4_lomap_desc::num_lomap) */
     for (size_t i = 0; i < aStrLocalMappings.size(); ++i)
     {
         com::Utf8Str strMapping(aStrLocalMappings[i]);
